refactor: move app to ouside of components and insert router
