--- client/src/components/App/PaddedContainer.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-import { Container } from 'semantic-ui-react';
-
-const PaddedContainer = styled(Container)`
-  padding-top: 2.5vw;
-  padding-bottom: .5vh;
-`;
-
-export default PaddedContainer;
